requirements.txt
cachetools==4.2.4
click==8.1.7
fastapi==0.68.0
h11==0.14.0
pydantic==1.8.2
PyJWT==2.8.0
starlette==0.14.2
typing_extensions==4.7.1
uvicorn==0.23.2
